requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
google-genai>=0.1.0
trafilatura>=1.6.0
